+++ R/portfolioBacktest.R
@@ -6,7 +6,7 @@
 #' function that takes as input a window of the stock prices and outputs the 
 #' portfolio weights. Multiple portfolios can be easily specified as a list 
 #' of functions or as files in a folder. Multiple datasets can be conveniently 
-#' obtained with the function \code{\link{stockDataResample}} that resamples
+#' obtained with the function \code{\link{financialDataResample}} that resamples
 #' the data downloaded with the function \code{\link{stockDataDownload}}.
 #' The results can be later assessed and arranged with tables and plots.
 #' The backtesting can be highly time-consuming depending on the number of 
@@ -25,15 +25,22 @@
 #'                       \href{https://CRAN.R-project.org/package=portfolioBacktest/vignettes/PortfolioBacktest.html}{vignette}
 #'                       for details.
 #' @param dataset_list List of datasets, each containing a list of \code{xts} objects, as generated
-#'                     by the function \code{\link{stockDataResample}}.
+#'                     by the function \code{\link{financialDataResample}}.
 #' @param folder_path If \code{portfolio_funs} is not defined, this should contain the path to a folder 
 #'                    containing the portfolio functions saved in files. See 
 #'                    \href{https://CRAN.R-project.org/package=portfolioBacktest/vignettes/PortfolioBacktest.html}{vignette}
 #'                    for details.
+#' @param source_to_local Logical value indicating whether to source files to local environment (default is \code{TRUE}).
+#'                        It might be dangerous to set it to \code{FALSE} as in such case the global environment may be changed.
+#'                        We suggest only to allow \code{FALSE} when the code in the source files does not work when locally 
+#'                        sourced, e.g., with some versions of package \code{CVXR}. In that case, we further recommend to set 
+#'                        \code{paral_portfolios > 1} to avoid changing the global environment.
 #' @param price_name Name of the \code{xts} column in each dataset that contains the prices to be used in the portfolio return 
 #'                   computation (default is \code{"adjusted"}).
-#' @param paral_portfolios Interger indicating number of portfolios to be evaluated in parallel (default is \code{1}).
-#' @param paral_datasets Interger indicating number of datasets to be evaluated in parallel (default is \code{1}).
+#' @param paral_portfolios Interger indicating number of portfolios to be evaluated in parallel (default is \code{1}),
+#'                         see \href{https://CRAN.R-project.org/package=portfolioBacktest/vignettes/PortfolioBacktest.html#parallel-backtesting}{vignette-paralle-mode} for details.
+#' @param paral_datasets Interger indicating number of datasets to be evaluated in parallel (default is \code{1}),
+#'                        see \href{https://CRAN.R-project.org/package=portfolioBacktest/vignettes/PortfolioBacktest.html#parallel-backtesting}{vignette-paralle-mode} for details.
 #' @param show_progress_bar Logical value indicating whether to show progress bar (default is \code{FALSE}). 
 #' @param benchmark String vector indicating the benchmark portfolios to be incorporated, currently supports:
 #' \itemize{\item{\code{uniform} - the uniform portfolio, \eqn{w = [1/N, ..., 1/N]} with \eqn{N} be number of stocks}
@@ -46,10 +53,12 @@
 #' @param T_rolling_window Length of the lookback rolling window (default is \code{252}).
 #' @param optimize_every How often the portfolio is to be optimized (default is \code{20}).
 #' @param rebalance_every How often the portfolio is to be rebalanced (default is \code{1}).
-#' @param execution String that can be either \code{"same day"} (default) or \code{"next day"}. 
+#' @param bars_per_year Number of bars per year. By default it will be calculated automatically 
+#'                      (e.g., for daily data there are 252 bars per year).
+#' @param execution String that can be either \code{"same period"} (default) or \code{"next period"}. 
 #'                  At the rebalancing period \code{t}, the portfolio has used information up to (and including)
-#'                  period \code{t}. Same day execution means one can get into the position at that period \code{t},
-#'                  whereas the next day execution means that one can only get into the position the following day.
+#'                  period \code{t}. Same period execution means one can get into the position at that period \code{t},
+#'                  whereas the next period execution means that one can only get into the position the following period.
 #' @param cost List containing four different types of transaction costs (common for all assets) 
 #'             for buying, selling, shorting, and long leveraging. The default is 
 #'             \code{cost = list(buy = 0e-4, sell = 0e-4, short = 0e-4, long_leverage = 0e-4)}.
@@ -59,7 +68,7 @@
 #' @param return_portfolio Logical value indicating whether to return the portfolios (default is \code{TRUE}).
 #'                         Two portfolios are returned: \code{w_designed} is the designed portfolio at each
 #'                         given rebalancing period (using all the information up to and including that period,
-#'                         which can be executed either on the same day or the following day)
+#'                         which can be executed either on the same period or the following period)
 #'                         and \code{w_bop} is the "beginning-of-period" portfolio (i.e., at each period it contains 
 #'                         the weights held in the market in the previous period so that the portfolio return at 
 #'                         that period is just the product of the asset returns and \code{w_bop} at that period.)
@@ -76,7 +85,7 @@
 #' 
 #' @author Daniel P. Palomar and Rui Zhou
 #' 
-#' @seealso \code{\link{stockDataDownload}}, \code{\link{stockDataResample}},
+#' @seealso \code{\link{stockDataDownload}}, \code{\link{financialDataResample}},
 #'          \code{\link{backtestSelector}}, \code{\link{backtestTable}}, 
 #'          \code{\link{backtestBoxPlot}}, \code{\link{backtestLeaderboard}},
 #'          \code{\link{backtestSummary}}, \code{\link{summaryTable}}, \code{\link{summaryBarPlot}}.
@@ -105,17 +114,16 @@
 #' 
 #' @import xts
 #' @importFrom zoo index
-#' @importFrom doSNOW registerDoSNOW
-#' @importFrom foreach foreach %dopar%
-#' @importFrom snow makeCluster stopCluster clusterExport
+#' @importFrom pbapply pblapply pboptions
+#' @importFrom parallel makeCluster stopCluster clusterExport clusterEvalQ
 #' @importFrom utils object.size
 #' @export
-portfolioBacktest <- function(portfolio_funs = NULL, dataset_list, folder_path = NULL, price_name = "adjusted",
+portfolioBacktest <- function(portfolio_funs = NULL, dataset_list, folder_path = NULL, source_to_local = TRUE, price_name = "adjusted",
                               paral_portfolios = 1, paral_datasets = 1,
                               show_progress_bar = FALSE, benchmark = NULL, 
                               shortselling = TRUE, leverage = Inf,
-                              T_rolling_window = 252, optimize_every = 20, rebalance_every = 1, 
-                              execution = c("same day", "next day"), 
+                              T_rolling_window = 252, optimize_every = 20, rebalance_every = 1, bars_per_year = NA,
+                              execution = c("same period", "next period"), 
                               cost = list(buy = 0e-4, sell = 0e-4, short = 0e-4, long_leverage = 0e-4),
                               cpu_time_limit = Inf,
                               return_portfolio = TRUE, return_returns = TRUE) {
@@ -129,21 +137,29 @@ portfolioBacktest <- function(portfolio_funs = NULL, dataset_list, folder_path =
   if (length(cost) != 4) stop("Problem in specifying the cost: the elements can only be buy, sell, short, and long_leverage.")
   if (is.xts(dataset_list[[1]])) stop("Each element of \"dataset_list\" must be a list of xts objects. Try to surround your passed \"dataset_list\" with list().")
   if (!(price_name %in% names(dataset_list[[1]]))) stop("Price data xts element \"", price_name, "\" does not exist in dataset_list.")
+  if (!is.xts(dataset_list[[1]][[1]])) stop("prices have to be xts.")
   ##############################
+  
+  if (!show_progress_bar) {
+    pbo <- pboptions(type = "none")
+    on.exit(pboptions(pbo), add = TRUE)
+  }
   
   # when portfolio_funs is passed
   if (!is.null(portfolio_funs)) {
     if (is.null(names(portfolio_funs))) portfolio_names <- paste0("fun", 1:length(portfolio_funs))
     else portfolio_names <- names(portfolio_funs)
-    
     if ("" %in% portfolio_names) stop("Each element of \"portfolio_funs\" must has a unique name.")
     if (length(portfolio_names) != length(unique(portfolio_names))) stop("\"portfolio_funs\" contains repeated names.")
+    
+    message(sprintf("\n [Backtesting %d portfolios over %d datasets (periodicity = %s data)]", 
+                    length(portfolio_names), length(dataset_list), periodicity(dataset_list[[1]][[1]])$scale))
     
     # in case the extra packages are loaded
     safeEvalPortf <- function(portfolio_fun, dataset_list, price_name,
                               paral_datasets, show_progress_bar,
                               shortselling, leverage,
-                              T_rolling_window, optimize_every, rebalance_every, 
+                              T_rolling_window, optimize_every, rebalance_every, bars_per_year,
                               execution, cost,
                               cpu_time_limit,
                               return_portfolio, return_returns) {
@@ -151,7 +167,7 @@ portfolioBacktest <- function(portfolio_funs = NULL, dataset_list, folder_path =
       res <- singlePortfolioBacktest(portfolio_fun, dataset_list, price_name, market = FALSE,
                                      paral_datasets, show_progress_bar,
                                      shortselling, leverage,
-                                     T_rolling_window, optimize_every, rebalance_every, 
+                                     T_rolling_window, optimize_every, rebalance_every, bars_per_year,
                                      execution, cost,
                                      cpu_time_limit,
                                      return_portfolio, return_returns)
@@ -169,40 +185,39 @@ portfolioBacktest <- function(portfolio_funs = NULL, dataset_list, folder_path =
         result[[i]] <- safeEvalPortf(portfolio_funs[[i]], dataset_list, price_name,
                                      paral_datasets, show_progress_bar,
                                      shortselling, leverage,
-                                     T_rolling_window, optimize_every, rebalance_every, 
+                                     T_rolling_window, optimize_every, rebalance_every, bars_per_year,
                                      execution, cost,
                                      cpu_time_limit,
                                      return_portfolio, return_returns)
       }
     } else {
-      # create the progress bar based on function number
-      if (show_progress_bar) {
-        sink(file = tempfile())
-        pb <- utils::txtProgressBar(max = length(portfolio_funs), style = 3)
-        sink()
-        opts <- list(progress = function(n) utils::setTxtProgressBar(pb, n))
-        message("Evaluating overall ", length(portfolio_funs), " portfolio functions in parallel")
-        opts$progress(0)
-      } else opts <- list()
       
+      # creata the cluster
       cl <- makeCluster(paral_portfolios)
-      registerDoSNOW(cl)
+      
+      # export global variables
       exports <- ls(envir = .GlobalEnv)
       exports <- exports[! exports %in% c("portfolio_fun", "dataset_list", "show_progress_bar")]
       large_objs <- exports[sapply(exports, function(x) object.size(get(x, envir = .GlobalEnv))) > 10 * 1024^2]  # filter the big objects more than 10 Mb
       exports <- exports[! exports %in% large_objs]
-      clusterExport(cl = cl, list = exports, envir = .GlobalEnv)
+      clusterExport(cl = cl, varlist = exports, envir = .GlobalEnv)
+      
+      # export attached packages
+      attached_packages <- .packages()
+      clusterExport(cl = cl, varlist = "attached_packages", envir = environment())
+      invisible(clusterEvalQ(cl = cl, expr = sapply(attached_packages, function(pkg) require(pkg, character.only = TRUE))))
+      
+      # apply!
       portfolio_fun <- NULL  # ugly hack to deal with CRAN note
-      result <- foreach(portfolio_fun = portfolio_funs, .combine = c, .packages = .packages(), .options.snow = opts) %dopar% {
-        return(list(safeEvalPortf(portfolio_fun, dataset_list, price_name,
-                                  paral_datasets, show_progress_bar,
-                                  shortselling, leverage,
-                                  T_rolling_window, optimize_every, rebalance_every, 
-                                  execution, cost,
-                                  cpu_time_limit,
-                                  return_portfolio, return_returns)))
-      }
-      if (show_progress_bar) close(pb)
+      result <- pblapply(portfolio_funs, 
+                         function(portfolio_fun) safeEvalPortf(portfolio_fun, dataset_list, price_name,
+                                                               paral_datasets, show_progress_bar,
+                                                               shortselling, leverage,
+                                                               T_rolling_window, optimize_every, rebalance_every, bars_per_year,
+                                                               execution, cost,
+                                                               cpu_time_limit,
+                                                               return_portfolio, return_returns), cl = cl)
+      # if (show_progress_bar) close(pb)
       stopCluster(cl) 
     }
     
@@ -216,18 +231,18 @@ portfolioBacktest <- function(portfolio_funs = NULL, dataset_list, folder_path =
     safeEvalFolder <- function(folder_path, file, dataset_list__, price_name,
                                paral_datasets, show_progress_bar,
                                shortselling, leverage,
-                               T_rolling_window, optimize_every, rebalance_every, 
+                               T_rolling_window, optimize_every, rebalance_every, bars_per_year,
                                execution, cost,
                                cpu_time_limit,
                                return_portfolio, return_returns) {
       packages_default <- search()  # snap the default packages
       
       error_message <- tryCatch({
-        suppressMessages(source(file.path(folder_path, file), local = TRUE))
+        suppressMessages(source(file.path(folder_path, file), local = source_to_local))
         res <- singlePortfolioBacktest(portfolio_fun, dataset_list__, price_name, market = FALSE,
                                        paral_datasets, show_progress_bar,
                                        shortselling, leverage,
-                                       T_rolling_window, optimize_every, rebalance_every, 
+                                       T_rolling_window, optimize_every, rebalance_every, bars_per_year,
                                        execution, cost,
                                        cpu_time_limit,
                                        return_portfolio, return_returns)
@@ -250,35 +265,23 @@ portfolioBacktest <- function(portfolio_funs = NULL, dataset_list, folder_path =
         result[[i]] <- safeEvalFolder(folder_path, files[i], dataset_list__, price_name,
                                       paral_datasets, show_progress_bar,
                                       shortselling, leverage,
-                                      T_rolling_window, optimize_every, rebalance_every, 
+                                      T_rolling_window, optimize_every, rebalance_every, bars_per_year,
                                       execution, cost,
                                       cpu_time_limit,
                                       return_portfolio, return_returns)
       }
     } else {
-      # creat the progress bar based on files number
-      if (show_progress_bar) {
-        sink(file = tempfile())
-        pb <- utils::txtProgressBar(max = length(files), style = 3)
-        sink()
-        opts <- list(progress = function(n) utils::setTxtProgressBar(pb, n))
-        message("Evaluating overall ", length(files), " portfolio functions (from files) in parallel")
-        opts$progress(0)
-      } else opts <- list()
       
       cl <- makeCluster(paral_portfolios)
-      registerDoSNOW(cl)
       file <- NULL  # ugly hack to deal with CRAN note
-      result <- foreach(file = files, .combine = c, .options.snow = opts) %dopar% {
-        return(list(safeEvalFolder(folder_path, file, dataset_list__, price_name,
-                                   paral_datasets, show_progress_bar,
-                                   shortselling, leverage,
-                                   T_rolling_window, optimize_every, rebalance_every, 
-                                   execution, cost,
-                                   cpu_time_limit,
-                                   return_portfolio, return_returns)))
-      }
-      if (show_progress_bar) close(pb)
+      result <- pblapply(files, function(file) safeEvalFolder(folder_path, file, dataset_list__, price_name,
+                                                              paral_datasets, show_progress_bar,
+                                                              shortselling, leverage,
+                                                              T_rolling_window, optimize_every, rebalance_every, bars_per_year,
+                                                              execution, cost,
+                                                              cpu_time_limit,
+                                                              return_portfolio, return_returns), cl = cl)
+      # if (show_progress_bar) close(pb)
       stopCluster(cl) 
     }
   }
@@ -289,7 +292,7 @@ portfolioBacktest <- function(portfolio_funs = NULL, dataset_list, folder_path =
   res_benchmark <- benchmarkBacktest(dataset_list, benchmark, price_name,
                                      paral_datasets, show_progress_bar,
                                      shortselling, leverage,
-                                     T_rolling_window, optimize_every, rebalance_every, 
+                                     T_rolling_window, optimize_every, rebalance_every, bars_per_year,
                                      execution, cost,
                                      cpu_time_limit,
                                      return_portfolio, return_returns)
@@ -306,7 +309,7 @@ portfolioBacktest <- function(portfolio_funs = NULL, dataset_list, folder_path =
 benchmarkBacktest <- function(dataset_list, benchmark, price_name,
                               paral_datasets, show_progress_bar,
                               shortselling, leverage,
-                              T_rolling_window, optimize_every, rebalance_every, 
+                              T_rolling_window, optimize_every, rebalance_every, bars_per_year,
                               execution, cost,
                               cpu_time_limit,
                               return_portfolio, return_returns) {
@@ -319,7 +322,7 @@ benchmarkBacktest <- function(dataset_list, benchmark, price_name,
     res$index <- singlePortfolioBacktest(portfolio_fun = NULL, dataset_list, price_name, market = TRUE,
                                          paral_datasets, show_progress_bar,
                                          shortselling, leverage,
-                                         T_rolling_window, optimize_every, rebalance_every, 
+                                         T_rolling_window, optimize_every, rebalance_every, bars_per_year,
                                          execution, cost,
                                          cpu_time_limit,
                                          return_portfolio, return_returns)
@@ -328,10 +331,11 @@ benchmarkBacktest <- function(dataset_list, benchmark, price_name,
   benchmark <- benchmark[benchmark != "index"]
   benchmark_portfolios <- benchmark_library[names(benchmark_library) %in% benchmark]
   if (length(benchmark_portfolios) > 0) {
-    res <- c(res, portfolioBacktest(benchmark_portfolios, dataset_list, NULL, price_name,
+    dataset_list <- lapply(dataset_list, function(x) {if (!("adjusted" %in% names(x))) x$adjusted <- x[[price_name]]; return(x)})
+    res <- c(res, portfolioBacktest(benchmark_portfolios, dataset_list, NULL, TRUE, price_name,
                                     1, paral_datasets, show_progress_bar, NULL,
                                     shortselling, leverage,
-                                    T_rolling_window, optimize_every, rebalance_every,
+                                    T_rolling_window, optimize_every, rebalance_every, bars_per_year,
                                     execution, cost,
                                     cpu_time_limit,
                                     return_portfolio, return_returns))
@@ -345,48 +349,43 @@ benchmarkBacktest <- function(dataset_list, benchmark, price_name,
 singlePortfolioBacktest <- function(portfolio_fun, dataset_list, price_name, market,
                                     paral_datasets, show_progress_bar,
                                     shortselling, leverage,
-                                    T_rolling_window, optimize_every, rebalance_every, 
+                                    T_rolling_window, optimize_every, rebalance_every, bars_per_year,
                                     execution, cost,
                                     cpu_time_limit,
                                     return_portfolio, return_returns) {
-  # create the progress bar
-  if (show_progress_bar) {
-    sink(file = tempfile())
-    pb <- utils::txtProgressBar(max = length(dataset_list), style = 3)
-    sink()
-    opts <- list(progress = function(n) utils::setTxtProgressBar(pb, n))
-    opts$progress(0)
-  } else opts <- list()
   
   # when price is a list of xts object
   if (paral_datasets == 1) { ########## no-parallel mode
     result <- list()
-    for (i in 1:length(dataset_list)) {
-      result[[i]] <- singlePortfolioSingleXTSBacktest(portfolio_fun, dataset_list[[i]], price_name, market,
-                                                      shortselling, leverage,
-                                                      T_rolling_window, optimize_every, rebalance_every, 
-                                                      execution, cost,
-                                                      cpu_time_limit,
-                                                      return_portfolio, return_returns)
-      if (show_progress_bar) opts$progress(i)
-    }
+    result <- pblapply(dataset_list, function(dat) singlePortfolioSingleXTSBacktest(portfolio_fun, dat, price_name, market,
+                                                                                    shortselling, leverage,
+                                                                                    T_rolling_window, optimize_every, rebalance_every, bars_per_year,
+                                                                                    execution, cost,
+                                                                                    cpu_time_limit,
+                                                                                    return_portfolio, return_returns))
   } else {               ########### parallel mode
     cl <- makeCluster(paral_datasets)
-    registerDoSNOW(cl)
+    # registerDoSNOW(cl)
+    
+    # export global variables
     exports <- ls(envir = .GlobalEnv)
     exports <- exports[! exports %in% c("portfolio_fun", "dat")]
     large_objs <- exports[sapply(exports, function(x) object.size(get(x, envir = .GlobalEnv))) > 10 * 1024^2]  # filter the big objects more than 10 Mb
     exports <- exports[! exports %in% large_objs]
-    clusterExport(cl = cl, list = exports, envir = .GlobalEnv)
+    clusterExport(cl = cl, varlist = exports, envir = .GlobalEnv)
+    
+    # export attached packages
+    attached_packages <- .packages()
+    clusterExport(cl = cl, varlist = "attached_packages", envir = environment())
+    invisible(clusterEvalQ(cl = cl, expr = sapply(attached_packages, function(pkg) require(pkg, character.only = TRUE))))
+    
     dat <- NULL  # ugly hack to deal with CRAN note
-    result <- foreach(dat = dataset_list, .combine = c, .packages = .packages(), .options.snow = opts) %dopar% {
-      return(list(singlePortfolioSingleXTSBacktest(portfolio_fun, dat, price_name, market,
-                                                   shortselling, leverage,
-                                                   T_rolling_window, optimize_every, rebalance_every, 
-                                                   execution, cost,
-                                                   cpu_time_limit,
-                                                   return_portfolio, return_returns)))
-    }
+    result <- pblapply(dataset_list, function(dat) singlePortfolioSingleXTSBacktest(portfolio_fun, dat, price_name, market,
+                                                                                    shortselling, leverage,
+                                                                                    T_rolling_window, optimize_every, rebalance_every, bars_per_year,
+                                                                                    execution, cost,
+                                                                                    cpu_time_limit,
+                                                                                    return_portfolio, return_returns), cl = cl)
     stopCluster(cl) 
   }
   
@@ -408,7 +407,7 @@ singlePortfolioBacktest <- function(portfolio_fun, dataset_list, price_name, mar
 #' @import xts
 singlePortfolioSingleXTSBacktest <- function(portfolio_fun, data, price_name, market,
                                              shortselling, leverage,
-                                             T_rolling_window, optimize_every, rebalance_every, 
+                                             T_rolling_window, optimize_every, rebalance_every, bars_per_year,
                                              execution, cost,
                                              cpu_time_limit,
                                              return_portfolio, return_returns) {
@@ -449,7 +448,7 @@ singlePortfolioSingleXTSBacktest <- function(portfolio_fun, data, price_name, ma
   if (optimize_every%%rebalance_every != 0) stop("The reoptimization period has to be a multiple of the rebalancing period.")
   if (anyNA(prices)) stop("prices contain NAs.")
   if (!is.function(portfolio_fun)) stop("portfolio_fun is not a function.")
-  if (periodicity(prices)$scale != "daily") stop("This function only accepts daily data.")
+  #if (periodicity(prices)$scale != "daily") stop("This function only accepts daily data.")
   #if (tzone(prices) != Sys.timezone()) tzone(prices) <- Sys.timezone()
   if (tzone(prices) != "UTC") tzone(prices) <- "UTC"
   #################################
@@ -464,18 +463,16 @@ singlePortfolioSingleXTSBacktest <- function(portfolio_fun, data, price_name, ma
   error <- flag_timeout <- FALSE; error_message <- NA; error_capture <- NULL; cpu_time <- c(); 
   w <- xts(matrix(NA, length(rebalance_indices), N), order.by = index(prices)[rebalance_indices])
   colnames(w) <- gsub(".Adjusted", "", colnames(prices))
-  
   for (i in 1:length(rebalance_indices)) {
-    
     idx_prices <- rebalance_indices[i]
     
     if (idx_prices %in% optimize_indices) {  # reoptimize
-      data_window  <- lapply(data, function(x) {x[(idx_prices-T_rolling_window+1):idx_prices, ]})
+      data_window  <- lapply(data, function(x) x[(idx_prices-T_rolling_window+1):idx_prices, ])
       start_time <- proc.time()[3] 
       error_capture <- R.utils::withTimeout(expr = evaluate::try_capture_stack(w[i, ] <- do.call(portfolio_fun, list(data_window)), environment()), 
                                             timeout = cpu_time_limit, onTimeout = "silent")
       cpu_time <- c(cpu_time, as.numeric(proc.time()[3] - start_time))
-    } else {# just rebalance without reoptimizing
+    } else { # just rebalance without reoptimizing
       w[i, ] <- w[i-1, ]
     }
     
@@ -519,7 +516,7 @@ singlePortfolioSingleXTSBacktest <- function(portfolio_fun, data, price_name, ma
   # all.equal(portf$rets, res_check$returns)
   # all.equal(portf$w_bop, res_check$BOP.Weight, check.attributes = FALSE)
 
-  res$performance <- portfolioPerformance(rets = portf$rets, ROT_bips = portf$ROT_bips)
+  res$performance <- portfolioPerformance(rets = portf$rets, ROT_bips = portf$ROT_bips, bars_per_year)
   res$cpu_time <- mean(cpu_time)
   res$error <- error
   res$error_message <- error_message
@@ -534,6 +531,8 @@ singlePortfolioSingleXTSBacktest <- function(portfolio_fun, data, price_name, ma
 
   return(res)
 }
+
+
 
 
 #' @title Add a new performance measure to backtests
@@ -602,7 +601,7 @@ add_performance <- function(bt, name, fun, desired_direction = 1) {
 #   rets: is an xts recording portfolio's return
 #   ROT_bips
 #
-portfolioPerformance <- function(rets = NA, ROT_bips = NA) {
+portfolioPerformance <- function(rets = NA, ROT_bips = NA, bars_per_year = NA) {
   performance <- rep(NA, 9)
   names(performance) <- c("Sharpe ratio", "max drawdown", "annual return", "annual volatility", 
                           "Sterling ratio", "Omega ratio", "ROT (bps)", "VaR (0.95)", "CVaR (0.95)")
@@ -610,16 +609,22 @@ portfolioPerformance <- function(rets = NA, ROT_bips = NA) {
   attr(performance, "judge") <- c(1, -1, 1, -1, 1, 1, 1, -1, -1)
   
   if (!anyNA(rets)) {
-    # fill the elements one by one
-    performance["Sharpe ratio"]      <- PerformanceAnalytics::SharpeRatio.annualized(rets)
-    performance["max drawdown"]      <- PerformanceAnalytics::maxDrawdown(rets)
-    performance["annual return"]     <- PerformanceAnalytics::Return.annualized(rets)  # prod(1 + rets)^(252/nrow(rets)) - 1 or mean(rets) * 252
-    performance["annual volatility"] <- PerformanceAnalytics::StdDev.annualized(rets)  # sqrt(252) * sd(rets, na.rm = TRUE)
-    performance["Sterling ratio"]    <- PerformanceAnalytics::Return.annualized(rets) / PerformanceAnalytics::maxDrawdown(rets)
-    performance["Omega ratio"]       <- PerformanceAnalytics::Omega(rets)
-    performance["ROT (bps)"]         <- ROT_bips
-    performance["VaR (0.95)"]        <- PerformanceAnalytics::VaR(rets, 0.95, method = "historical", invert = FALSE)
-    performance["CVaR (0.95)"]       <- PerformanceAnalytics::CVaR(rets, 0.95, method = "historical", invert = FALSE)
+    fraction_in <- sum(abs(rets) > 1e-8)/nrow(rets)
+    rets <- rets[abs(rets) > 1e-8]  # remove data where return is zero
+    if (nrow(rets) == 0)
+      performance[] <- c(NA, 0, 0, 0, NA, NA, NA, 0, 0)
+    else {
+      # fill the elements one by one
+      performance["Sharpe ratio"]      <- PerformanceAnalytics::SharpeRatio.annualized(rets, scale = bars_per_year) * sqrt(fraction_in)
+      performance["max drawdown"]      <- PerformanceAnalytics::maxDrawdown(rets)
+      performance["annual return"]     <- PerformanceAnalytics::Return.annualized(rets, scale = bars_per_year) * fraction_in        # prod(1 + rets)^(252/nrow(rets)) - 1 or mean(rets) * 252
+      performance["annual volatility"] <- PerformanceAnalytics::StdDev.annualized(rets, scale = bars_per_year) * sqrt(fraction_in)  # sqrt(252) * sd(rets, na.rm = TRUE)
+      performance["Sterling ratio"]    <- performance["annual return"] / performance["max drawdown"]
+      performance["Omega ratio"]       <- PerformanceAnalytics::Omega(rets)
+      performance["ROT (bps)"]         <- ROT_bips
+      performance["VaR (0.95)"]        <- PerformanceAnalytics::VaR(rets, 0.95, method = "historical", invert = FALSE)
+      performance["CVaR (0.95)"]       <- PerformanceAnalytics::CVaR(rets, 0.95, method = "historical", invert = FALSE)
+    }
   }
   return(performance)
 }
@@ -630,19 +635,18 @@ portfolioPerformance <- function(rets = NA, ROT_bips = NA) {
 # Computes the returns of a portfolio of several assets (ignoring transaction costs):
 #   R: is an xts with the individual asset linear returns (not log returns)
 #   weights: is an xts with the normalized dollar allocation (wrt NAV, typically with sum=1) where
-#            - each row represents a rebalancing date (with portfolio computed with info up to and including that day)
+#            - each row represents a rebalancing date (with portfolio computed with info up to and including that period)
 #            - dates with no rows means: no rebalancing (note that the portfolio may then violate some margin constraints...)
-#   rets: are the returns for each day at the close
+#   rets: are the returns for each period at the close
 #
 #' @import xts
 returnPortfolio <- function(R, weights, 
-                            execution = c("same day", "next day"), 
+                            execution = c("same period", "next period"), 
                             cost = list(buy = 0*10^(-4), sell = 0*10^(-4), short = 0*10^(-4), long_leverage = 0*10^(-4)),
                             initial_cash = 1) {
   ######## error control  #########
   if (!is.xts(R) || !is.xts(weights)) stop("This function only accepts xts.")
-  if (periodicity(R)$scale != "daily")
-    stop("This function only accepts daily data.")
+  # if (periodicity(R)$scale != "daily") stop("This function only accepts daily data.")
   if (!all(index(weights) %in% index(R))) stop("Weight dates do not appear in the returns")
   if (ncol(R) != ncol(weights)) stop("Number of weights does not match the number of assets in the returns")
   if (anyNA(R[-1])) stop("Returns contain NAs")
@@ -657,12 +661,12 @@ returnPortfolio <- function(R, weights,
   w <- R; w[] <- NA; colnames(w) <- colnames(weights)
   w[as.character(index(weights)), ] <- weights
   w <- switch(match.arg(execution),  # w[t] used info up to (including) price[t]
-              "same day" = lag.xts(w, 1),  # w[t] is (idealistically) executed at price[t], so will multiply return[t+1]
-              "next day" = lag.xts(w, 2),  # w[t] is executed one period later at price[t+1], so will multiply return[t+2]
+              "same period" = lag.xts(w, 1),  # w[t] is (idealistically) executed at price[t], so will multiply return[t+1]
+              "next period" = lag.xts(w, 2),  # w[t] is executed one period later at price[t+1], so will multiply return[t+2]
               stop("Execution method unknown"))
   after_rebalance_indices <- which(!is.na(w[, 1]))
-  # loop    (NAV contains the NAV at the beginning of the day before the open (i.e., end of previous day), like w,
-  #          whereas ret contains the returns at the end of the day, so lag(ret) equals (NAV - lag(NAV))/lag(NAV))
+  # loop    (NAV contains the NAV at the beginning of the period before the open (i.e., end of previous period), like w,
+  #          whereas ret contains the returns at the end of the period, so lag(ret) equals (NAV - lag(NAV))/lag(NAV))
   NAV <- ret <- xts(rep(NA, nrow(R)), order.by = index(R))
   colnames(ret) <- "portfolio return"
   colnames(NAV) <- "NAV"
@@ -675,7 +679,7 @@ returnPortfolio <- function(R, weights,
     if (t %in% after_rebalance_indices) {
       delta_rel[t, ] <- w[t, ] - w_eop
       if (compute_tc)
-        tc <- cost$buy*sum(pos(delta_rel[t, ])) + cost$sell*sum(pos(-delta_rel[t, ])) +  # trading cost
+        tc <- cost$buy*sum(pos(delta_rel[t, ])) + cost$sell*sum(pos(-delta_rel[t, ])) +       # trading cost
               cost$long_leverage*max(0, sum(pos(w[t, ])) - 1) + cost$short*sum(pos(-w[t, ]))  # borrowing cost
       cash <- 1 - sum(w[t, ]) - tc  # normalized cash wrt NAV[t]
       ret[t] <- sum(R[t, ]*w[t, ]) - tc  # recall w is normalized wrt NAV[t]